MCP - Nuke resources folder, fold all into tools
We don't use MCP resources anymore — everything is registered via
mcp.tool() now (opencode doesn't auto-pull resources into model context;
they only enter via the redundant `read_resource` wrapper). Folder name
no longer reflects what's inside.

- Move alpha_lab, delta_lab, discovery, tokens to tools/ (no collision).
- Merge resources/contracts.py contents into tools/contracts.py
  (contracts_list, contracts_get).
- Merge resources/hyperliquid.py contents into tools/hyperliquid.py
  (10 read funcs).
- Merge resources/wallets.py contents into tools/wallets.py
  (onchain_list_wallets, core_get_wallet, core_get_wallet_balances,
  onchain_get_wallet_activity).
- Drop the now-empty resources/ directory.
- Update server.py imports and four test imports.

Tool count unchanged (48). All affected tests pass.
