+++ wayfinder_paths/mcp/tools/alpha_lab.py
