+++ wayfinder_paths/mcp/tools/delta_lab.py
