+++ wayfinder_paths/mcp/tools/discovery.py
