+++ wayfinder_paths/tests/test_mcp_tokens.py
@@ -5,7 +5,7 @@ from unittest.mock import AsyncMock, patch
 
 import pytest
 
-from wayfinder_paths.mcp.resources.tokens import (
+from wayfinder_paths.mcp.tools.tokens import (
     onchain_fuzzy_search_tokens,
     onchain_get_gas_token,
     onchain_resolve_token,
@@ -17,7 +17,7 @@ async def test_resolve_token_happy_path():
     fake_client = AsyncMock()
     fake_client.get_token_details = AsyncMock(return_value={"symbol": "USDC"})
 
-    with patch("wayfinder_paths.mcp.resources.tokens.TOKEN_CLIENT", fake_client):
+    with patch("wayfinder_paths.mcp.tools.tokens.TOKEN_CLIENT", fake_client):
         out = await onchain_resolve_token("usd-coin-arbitrum")
 
     result = json.loads(out)
@@ -29,7 +29,7 @@ async def test_get_gas_token_happy_path():
     fake_client = AsyncMock()
     fake_client.get_gas_token = AsyncMock(return_value={"symbol": "ETH"})
 
-    with patch("wayfinder_paths.mcp.resources.tokens.TOKEN_CLIENT", fake_client):
+    with patch("wayfinder_paths.mcp.tools.tokens.TOKEN_CLIENT", fake_client):
         out = await onchain_get_gas_token("arbitrum")
 
     result = json.loads(out)
@@ -41,7 +41,7 @@ async def test_fuzzy_search_tokens_happy_path():
     fake_client = AsyncMock()
     fake_client.fuzzy_search = AsyncMock(return_value={"results": [{"id": "foo"}]})
 
-    with patch("wayfinder_paths.mcp.resources.tokens.TOKEN_CLIENT", fake_client):
+    with patch("wayfinder_paths.mcp.tools.tokens.TOKEN_CLIENT", fake_client):
         out = await onchain_fuzzy_search_tokens(chain_code="arbitrum", query="usd")
 
     result = json.loads(out)
